requirements.txt
aiosqlite==0.21.0
altgraph==0.17.4
amqp==5.3.1
asgiref==3.9.0
billiard==4.2.1
celery==5.3.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
distlib==0.3.9
Django==4.2.7
django-celery-beat==2.8.1
django-cors-headers==4.8.0
django-crontab==0.7.1
django-filter==25.1
django-ratelimit==4.1.0
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
docutils==0.21.2
drf-yasg==1.21.10
filelock==3.18.0
filetype==1.2.0
gem==0.1.12
graphene==3.4.3
graphene-django==3.1.5
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
joblib==1.5.1
kombu==5.5.4
Markdown==3.8.2
nltk==3.9.1
packaging==25.0
parameterized==0.9.0
pefile==2023.2.7
pillow==10.4.0
pipenv==2025.0.3
platformdirs==4.3.8
promise==2.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.7
PyJWT==2.10.1
PyPDF2==3.0.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytube3==9.6.4
pytz==2025.2
PyYAML==6.0.2
redis==4.6.0
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.23
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
python-decouple==3.8
